Rename and combine some files
