+++ src/helpers/message-helper.ts
@@ -6,19 +6,16 @@
 /* global document */
 
 export function showMessage(text: string): void {
-  //document.getElementById(".welcome-body").style.display = "none";
   document.getElementById("message-area").style.display = "flex";
   document.getElementById("message-area").innerText = text;
 }
 
 export function clearMessage(): void {
-  //document.getElementById(".welcome-body").style.display = "none";
   document.getElementById("message-area").style.display = "flex";
   document.getElementById("message-area").innerText = "---<br>";
 }
 
 export function hideMessage(): void {
-  //document.getElementById(".welcome-body").style.display = "flex";
   document.getElementById("message-area").style.display = "none";
   document.getElementById("message-area").innerText = "---<br>";
 }
